Remove utils
