+++ beacon_node/eth1_chain/src/web3_fetcher.rs
@@ -1,3 +1,5 @@
+use bls::{PublicKeyBytes, SignatureBytes};
+use ethabi::{decode, ParamType, Token};
 use parking_lot::RwLock;
 use std::collections::BTreeMap;
 use std::sync::Arc;
@@ -10,7 +12,6 @@ use web3::types::*;
 use web3::Web3;
 
 use crate::fetcher::Eth1DataFetcher;
-use crate::utils::*;
 
 /// Config for an Eth1 chain contract.
 #[derive(Debug, Clone)]
@@ -128,6 +129,58 @@ impl Eth1DataFetcher for Web3DataFetcher {
             })
             .map_err(|_| ());
         Box::new(event_future)
+    }
+}
+
+// Converts a valid vector to a u64.
+pub fn vec_to_u64_le(bytes: &[u8]) -> Option<u64> {
+    let mut array = [0; 8];
+    if bytes.len() == 8 {
+        let bytes = &bytes[..array.len()];
+        array.copy_from_slice(bytes);
+        Some(u64::from_le_bytes(array))
+    } else {
+        None
+    }
+}
+
+/// Parse contract logs.
+pub fn parse_logs(log: Log, types: &[ParamType]) -> Option<Vec<Token>> {
+    decode(types, &log.data.0).ok()
+}
+
+/// Parse logs from deposit contract.
+pub fn parse_deposit_logs(log: Log) -> Option<(u64, DepositData)> {
+    let deposit_event_params = &[
+        ParamType::FixedBytes(48), // pubkey
+        ParamType::FixedBytes(32), // withdrawal_credentials
+        ParamType::FixedBytes(8),  // amount
+        ParamType::FixedBytes(96), // signature
+        ParamType::FixedBytes(8),  // index
+    ];
+    let parsed_logs = parse_logs(log, deposit_event_params).unwrap();
+    // Convert from tokens to Vec<u8>.
+    let params = parsed_logs
+        .into_iter()
+        .map(|x| match x {
+            Token::FixedBytes(v) => Some(v),
+            _ => None,
+        })
+        .collect::<Option<Vec<_>>>()?;
+
+    // Event should have exactly 5 parameters.
+    if params.len() == 5 {
+        Some((
+            vec_to_u64_le(&params[4])?,
+            DepositData {
+                pubkey: PublicKeyBytes::from_bytes(&params[0]).unwrap(),
+                withdrawal_credentials: H256::from_slice(&params[1]),
+                amount: vec_to_u64_le(&params[2])?,
+                signature: SignatureBytes::from_bytes(&params[3]).ok()?,
+            },
+        ))
+    } else {
+        None
     }
 }
